Restore backtrace support
