+++ CMakeLists.txt
@@ -123,6 +123,18 @@ set(NANO_ROCKSDB_TOOLS
     OFF
     CACHE BOOL "")
 
+option(NANO_STACKTRACE_BACKTRACE
+       "Use BOOST_STACKTRACE_USE_BACKTRACE in stacktraces, for POSIX" OFF)
+
+if(NANO_STACKTRACE_BACKTRACE)
+  add_definitions(-DNANO_STACKTRACE_BACKTRACE)
+  add_definitions(-DBOOST_STACKTRACE_USE_BACKTRACE)
+  if(NANO_BACKTRACE_INCLUDE)
+    add_definitions(
+      -DBOOST_STACKTRACE_BACKTRACE_INCLUDE_FILE=${NANO_BACKTRACE_INCLUDE})
+  endif()
+endif()
+
 if(${NANO_TIMED_LOCKS} GREATER 0)
   add_definitions(-DNANO_TIMED_LOCKS=${NANO_TIMED_LOCKS})
   add_definitions(-DNANO_TIMED_LOCKS_FILTER=${NANO_TIMED_LOCKS_FILTER})
